# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Zuultimate
  version: 1.0.0
paths:
...
  /v1/crm/configs:
    post:
      tags:
      - crm
      summary: Create CRM config
      operationId: create_config_v1_crm_configs_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CRMConfigCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CRMConfigResponse'
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
      x-stability: stable
      x-required-scopes:
      - identity:write
  /v1/crm/sync:
    post:
      tags:
      - crm
      summary: Start CRM sync job
      operationId: start_sync_v1_crm_sync_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncStartRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncJobResponse'
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
      x-stability: stable
      x-required-scopes:
      - identity:write
  /v1/crm/sync/{job_id}:
    get:
      tags:
      - crm
      summary: Get sync job status
      operationId: get_sync_status_v1_crm_sync__job_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncJobResponse'
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-stability: stable
      x-required-scopes:
      - identity:read
  /v1/crm/adapters:
    get:
      tags:
      - crm
      summary: List CRM adapters
      description: List all registered CRM provider adapters.
      operationId: available_adapters_v1_crm_adapters_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
      x-stability: stable
      x-required-scopes:
      - identity:read
  /v1/crm/adapters/{provider}/test:
    post:
      tags:
      - crm
      summary: Test CRM adapter connectivity
      description: Test connectivity to a CRM provider.
      operationId: test_adapter_v1_crm_adapters__provider__test_post
      security:
      - HTTPBearer: []
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          title: Provider
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-stability: stable
      x-required-scopes:
      - identity:write
  /v1/crm/adapters/{provider}/fetch:
    post:
      tags:
      - crm
      summary: Fetch contacts from adapter
      description: Fetch sample contacts from a CRM provider adapter.
      operationId: fetch_contacts_v1_crm_adapters__provider__fetch_post
      security:
      - HTTPBearer: []
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          title: Provider
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-stability: stable
      x-required-scopes:
      - identity:write
...
  /v1/plugins/{plugin_id}/webhook:
    post:
      tags:
      - plugins
      summary: Forward webhook to plugin
      description: Forward a webhook payload to the specified plugin.
      operationId: plugin_webhook_v1_plugins__plugin_id__webhook_post
      security:
      - HTTPBearer: []
      parameters:
      - name: plugin_id
        in: path
        required: true
        schema:
          type: string
          title: Plugin Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Plugin Webhook V1 Plugins  Plugin Id  Webhook Post
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-stability: stable
      x-required-scopes:
      - identity:write
  /v1/webhooks:
    get:
      tags:
      - webhooks
      summary: List all webhooks
      operationId: list_webhooks_v1_webhooks_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/WebhookResponse'
                type: array
                title: Response List Webhooks V1 Webhooks Get
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
      x-stability: stable
      x-required-scopes:
      - identity:read
    post:
      tags:
      - webhooks
      summary: Create webhook
      operationId: create_webhook_v1_webhooks_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookCreateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
      x-stability: stable
      x-required-scopes:
      - identity:write
  /v1/webhooks/{webhook_id}:
    delete:
      tags:
      - webhooks
      summary: Delete webhook
      operationId: delete_webhook_v1_webhooks__webhook_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: string
          title: Webhook Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-stability: stable
      x-required-scopes:
      - identity:write
...
components:
  schemas:
...
    CRMConfigCreate:
      properties:
        provider:
          type: string
          title: Provider
          description: CRM provider name
          examples:
          - salesforce
        api_url:
          type: string
          title: Api Url
          description: Provider API base URL
          default: ''
      type: object
      required:
      - provider
      title: CRMConfigCreate
    CRMConfigResponse:
      properties:
        id:
          type: string
          title: Id
          description: Config UUID
        provider:
          type: string
          title: Provider
          description: CRM provider name
        api_url:
          type: string
          title: Api Url
          description: Provider API base URL
        is_active:
          type: boolean
          title: Is Active
          description: Whether this config is active
      type: object
      required:
      - id
      - provider
      - api_url
      - is_active
      title: CRMConfigResponse
...
    ErrorResponse:
      properties:
        error:
          type: string
          title: Error
          description: Error message
        code:
          type: string
          title: Code
          description: Machine-readable error code
          default: ZUUL_ERROR
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
          description: Additional error context
      type: object
      required:
      - error
      title: ErrorResponse
...
    SyncJobResponse:
      properties:
        id:
          type: string
          title: Id
          description: Sync job UUID
        config_id:
          type: string
          title: Config Id
          description: CRM config UUID
        status:
          type: string
          title: Status
          description: Job status (pending/running/completed/failed)
        records_synced:
          type: integer
          title: Records Synced
          description: Number of records synced so far
      type: object
      required:
      - id
      - config_id
      - status
      - records_synced
      title: SyncJobResponse
    SyncStartRequest:
      properties:
        config_id:
          type: string
          title: Config Id
          description: CRM config UUID to sync from
      type: object
      required:
      - config_id
      title: SyncStartRequest
...
    WebhookCreateRequest:
      properties:
        url:
          type: string
          minLength: 1
          title: Url
        events_filter:
          type: string
          title: Events Filter
          default: '*'
        secret:
          type: string
          title: Secret
          default: ''
        description:
          type: string
          title: Description
          default: ''
      type: object
      required:
      - url
      title: WebhookCreateRequest
    WebhookResponse:
      properties:
        id:
          type: string
          title: Id
        url:
          type: string
          title: Url
        events_filter:
          type: string
          title: Events Filter
        is_active:
          type: boolean
          title: Is Active
        description:
          type: string
          title: Description
      type: object
      required:
      - id
      - url
      - events_filter
      - is_active
      - description
      title: WebhookResponse
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer